Add in craft item market pricing Craft lists can have world preferences for calculating the best price Fix potential memory retention New craft pricer service Add hosted universalis(to be eventually split into it's own lib)

+++ Time/BitfieldUptime.cs
@@ -65,7 +65,7 @@ public readonly struct BitfieldUptime : IEquatable<BitfieldUptime>
         if (hours == 0)
             return BadHour;
 
-        var rotatedHours = currentHour == 0 ? hours : (hours >> (int)currentHour) | ((hours << (32 - (int)currentHour)) >> 8);
+        var rotatedHours = currentHour == 0 ? hours : (hours >> currentHour) | ((hours << (32 - currentHour)) >> 8);
         return Bits.TrailingZeroCount(rotatedHours);
     }
